// src/codegen/cgconstants.ts
//src\app_front\codegen\cgconstants.ts

/**
 * CodeGen Formats
 */
export enum CodeFormats {
    TXT = "textplain",
    TYPESCRIPT = "typescript",
    SQL = "sql",
    PYTHON = "pyhton",
    JAVASCRIPT = "json",
    JSON = "json",
    MD = "mardown",
    CSS = "css",
    HTML = "html"
}

/**
 * CodeGen File Extensions
 */
export enum FilesExtensions {
    TXT = ".txt",   
    TYPESCRIPT = ".ts",
    TYPESCRIPT_JSX = ".tsx",
    SQL = ".sql",
    PYTHON = ".py",
    JAVASCRIPT = ".js",
    JSON = ".json",
    MD = ".md",
    CSS = ".css",
    HTML = ".html"
}